requirements.txt
aiogram==2.11.2
aiohttp==3.7.2
aioredis==1.3.1
arrow==0.17.0
asn1crypto==1.4.0
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.0
base58==2.0.1
bit==0.7.2
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==6.6
coincurve==13.0.0
docker-py==1.8.0
ecdsa==0.16.1
flake8==3.8.4
funcsigs==1.0.2
future==0.18.2
hiredis==1.1.0
idna==2.10
iniconfig==1.1.1
jsonrpcclient==2.0.1
jsonrpcserver==3.1.1
jsonschema==3.2.0
mccabe==0.6.1
mnemonic==0.19
multidict==5.0.2
packaging==20.4
path==15.0.0
path.py==12.5.0
pexpect==4.8.0
pluggy==0.13.1
protobuf==3.14.0
ptyprocess==0.6.0
py==1.9.0
pyaes==1.6.1
pyaml==20.4.0
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.9.9
pyflakes==2.2.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.1.2
python-binance-chain==0.1.20
python-dateutil==2.8.1
pytz==2020.4
pywallet==0.1.0
PyYAML==5.3.1
qrcode==6.1
requests==2.25.0
secp256k1==0.13.2
sha256==0.1
six==1.15.0
tabulate==0.8.7
toml==0.10.2
two1==3.10.9
typing-extensions==3.7.4.3
ujson==4.0.1
urllib3==1.26.2
websocket-client==0.57.0
websockets==8.1
yarl==1.6.3
